Change the base Dockerfile image to Ubuntu
The latest version of Alpine Linux 3.17 does not support Varnish 6.
So either the varnish needed to be build from source or the base image needed to be changed.
Ubuntu was a good substitute for Alpine Linux as it is used in the other varnish services.

The varnish-modules is bumped to 6.0-lts version. This is because the specific commit was not able to compile with 6.0-lts